+++ pgcli/magic.py
@@ -25,7 +25,11 @@ def pgcli_line_magic(line):
     if hasattr(sql.connection.Connection, "get"):
         conn = sql.connection.Connection.get(parsed["connection"])
     else:
-        conn = sql.connection.Connection.set(parsed["connection"])
+        try:
+            conn = sql.connection.Connection.set(parsed["connection"])
+        # a new positional argument was added to Connection.set in version 0.4.0 of ipython-sql
+        except TypeError:
+            conn = sql.connection.Connection.set(parsed["connection"], False)
 
     try:
         # A corresponding pgcli object already exists