+++ changelog.rst
@@ -14,6 +14,7 @@ Bug fixes:
 * Fix issue where `syntax_style` config value would not have any effect. (#1212)
 * Fix crash because of not found `InputMode.REPLACE_SINGLE` with prompt-toolkit < 3.0.6
 * Fix comments being lost in config when saving a named query. (#1240)
+* Fix IPython magic for ipython-sql >= 0.4.0
 
 3.1.0
 =====